// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/DiscountCodeDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DiscountCodeDto = {
    ai: number;
    codeName: string;
    discountType: string;
    amount: number;
    usageCount: number;
    startDate: string;
    endDate: string;
    isActive: boolean;
    isExpired: boolean;
    modified: string;
};